Refactor CircleCI repo onboarding into one file (#2806)
